Control flow with custom errors types
Major changes:

* Remove the concept of `ScanStatus` in favor of custom errors

Minor changes:

* Move optional config to new file

https://github.com/harlow/kinesis-consumer/issues/75
