+++ examples/consumer/cp-dynamo/main.go
@@ -54,7 +54,7 @@ func main() {
 	}
 
 	var (
-		app    = flag.String("app", "", "App name")
+		app    = flag.String("app", "", "Consumer app name")
 		stream = flag.String("stream", "", "Stream name")
 		table  = flag.String("table", "", "Checkpoint table name")
 	)
@@ -103,11 +103,9 @@ func main() {
 	}()
 
 	// scan stream
-	err = c.Scan(ctx, func(r *consumer.Record) consumer.ScanStatus {
+	err = c.Scan(ctx, func(r *consumer.Record) error {
 		fmt.Println(string(r.Data))
-
-		// continue scanning
-		return consumer.ScanStatus{}
+		return nil // continue scanning
 	})
 	if err != nil {
 		log.Log("scan error: %v", err)
