+++ examples/consumer/cp-postgres/main.go
@@ -15,7 +15,7 @@ import (
 
 func main() {
 	var (
-		app     = flag.String("app", "", "App name")
+		app     = flag.String("app", "", "Consumer app name")
 		stream  = flag.String("stream", "", "Stream name")
 		table   = flag.String("table", "", "Table name")
 		connStr = flag.String("connection", "", "Connection Str")
@@ -53,11 +53,9 @@ func main() {
 	}()
 
 	// scan stream
-	err = c.Scan(ctx, func(r *consumer.Record) consumer.ScanStatus {
+	err = c.Scan(ctx, func(r *consumer.Record) error {
 		fmt.Println(string(r.Data))
-
-		// continue scanning
-		return consumer.ScanStatus{}
+		return nil // continue scanning
 	})
 
 	if err != nil {
